Move grammar AST classes to a separate directory
