+++ MetaMark/src/generation/ElementNodeGenerator.ts
@@ -1,7 +1,7 @@
 import { NodeGenerator } from "./NodeGenerator";
-import { DefinitionTerm } from "../DefinitionTerm";
+import { DefinitionTerm } from "../metamark_ast/DefinitionTerm";
 import { PropertySignatureTerm } from "../typescript_ast";
-import { ElementDefinitionTerm } from "../ElementDefinitionTerm";
+import { ElementDefinitionTerm } from "../metamark_ast/ElementDefinitionTerm";
 
 export class ElementNodeGenerator extends NodeGenerator {
     readonly _elementDefinition: ElementDefinitionTerm;
