+++ MetaMark/src/metamark_ast/DefinitionTerm.ts
@@ -1,5 +1,5 @@
-import { ParseFunctionGenerator } from "./generation/ParseFunctionGenerator";
-import { NodeGenerator } from "./generation/NodeGenerator";
+import { ParseFunctionGenerator } from "../generation/ParseFunctionGenerator";
+import { NodeGenerator } from "../generation/NodeGenerator";
 
 export abstract class DefinitionTerm {
     abstract readonly name: string;
