+++ MetaMark/src/metamark_ast/ExpressionTerm.ts
@@ -1,4 +1,4 @@
-import * as typescript_ast from "./typescript_ast";
+import * as typescript_ast from "../typescript_ast";
 
 export abstract class ExpressionTerm {
     abstract generateParseFunctionExpression(): typescript_ast.ExpressionTerm;
