+++ MetaMark/src/metamark_ast/StaticScope.ts
