+++ MetaMark/src/metamark_ast/StringLiteralTerm.ts
@@ -1,4 +1,4 @@
-import { StringLiteralContext } from "../generated_src/MetamarkParser";
+import { StringLiteralContext } from "../../generated_src/MetamarkParser";
 
 export class StringLiteralTerm {
     static equals(a: StringLiteralTerm, b: StringLiteralTerm): boolean;
